requirements.txt
absl-py==0.7.1
airtest==1.0.26
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.5.0
Appium-Python-Client==0.40
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
baidu-aip==2.2.13.0
bandit==1.5.1
beautifulsoup4==4.7.1
bleach==3.1.0
cached-property==1.5.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
comtypes==1.1.7
constantly==15.1.0
crcmod==1.7
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
DBUtils==1.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.0.9
django-ckeditor==5.6.1
django-filter==2.1.0
django-js-asset==1.1.0
django-mdeditor==0.1.9
djangorestframework==3.9.2
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
EasyProcess==0.2.5
entrypoints==0.3
facebook-wda==0.2.1
flake8==3.7.4
Flask==1.0.2
freeze==1.0.10
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
grpcio==1.20.1
h5py==2.9.0
hrpc==1.0.8
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.3
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.8
numpy==1.16.3
opencv-contrib-python==3.4.2.17
oss2==2.6.1
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
pbr==5.1.2
pickleshare==0.7.5
Pillow==5.3.0
pipenv==2018.11.26
pocoui==1.0.76
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
pyaml==16.12.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.7.3
PyDispatcher==2.0.5
pyflakes==2.1.0
PyGithub==1.39
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.1.1
PyMySQL==0.9.2
pyOpenSSL==19.0.0
pyparsing==2.3.0
pypiwin32==223
pyquery==1.4.0
pytesseract==0.2.6
python-dateutil==2.7.5
python-gitlab==1.8.0
python-memcached==1.59
pytz==2018.7
PyVirtualDisplay==0.2.1
pywin32==224
pywinauto==0.6.3
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.3
queuelib==1.5.0
redis==3.2.0
requests==2.20.1
schedule==0.6.0
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.11.0
smmap2==2.0.5
soupsieve==1.7.3
stevedore==1.30.0
svn==0.3.44
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texttable==0.9.1
tornado==5.1.1
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.3.5
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.5.1
virtualenvwrapper-win==1.2.5
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.55.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
yapf==0.25.0
zope.interface==4.6.0
